Refactor api.go
