+++ api/metric/batch.go
@@ -1,0 +1,1 @@
+package metric
